增加严格编译检查

+++ src/kalmanFiltering.cpp
@@ -2,7 +2,7 @@
  * @Author: yao.xie 1595341200@qq.com
  * @Date: 2024-03-11 17:25:32
  * @LastEditors: yao.xie 1595341200@qq.com
- * @LastEditTime: 2024-04-12 15:56:22
+ * @LastEditTime: 2024-07-01 20:57:38
  * @FilePath: /cplusplus/submodule/algorithmBase/src/kalmanFiltering.cpp
  * @Description:
  *
@@ -27,24 +27,24 @@ KalmanFilter::KalmanFilter() {  // 动力学模型转移矩阵
     H.resize(4, 4);
     x.resize(4);
     //clang-format off
-    A << 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0;
+    A << 1.0f, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0;
 
     x << 0, 0, 0, 0;
 
-    Q << 0.01, 0, 0, 0, 0, 0.01, 0, 0, 0, 0, 0.01, 0, 0, 0, 0, 0.01;
+    Q << 0.01f, 0, 0, 0, 0, 0.01f, 0, 0, 0, 0, 0.01f, 0, 0, 0, 0, 0.01f;
 
-    P << 0.01, 0, 0, 0, 0, 0.01, 0, 0, 0, 0, 0.01, 0, 0, 0, 0, 0.01;
+    P << 0.01f, 0, 0, 0, 0, 0.01f, 0, 0, 0, 0, 0.01f, 0, 0, 0, 0, 0.01f;
 
-    R << 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1;
+    R << 1.0f, 0, 0, 0, 0, 1.0f, 0, 0, 0, 0, 1.0f, 0, 0, 0, 0, 1.0f;
 
-    H << 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1;
+    H << 1.0f, 0, 0, 0, 0, 1.0f, 0, 0, 0, 0, 1.0f, 0, 0, 0, 0, 1.0f;
     //clang-format on
 }
 
 void KalmanFilter::predict(double dt) {
     // 预测步骤
-    A(0, 2) = dt;
-    A(1, 3) = dt;
+    A(0, 2) = static_cast<float>(dt);
+    A(1, 3) = static_cast<float>(dt);
     x_hat = A * x;
     P_hat = A * P * A.transpose() + Q;
 }
